Need to split windows into two
Fix cs_window so it doesn't register itself: otherwise we can't create a
second window.

Track the ID of the focused window so we know which window is being
split. Put the focus on the new window.

TODO:
- Add border: how do we know which windows need borders?

Add atomic operations to IO: a window needs to set the cursor position
and write without being interrupted. To do this I created an "atomic
operations" capability.