More work on the window tree
See notes.

I'm trying to work out a few edge cases:

+---+---+  This is confusing to model.
| x |   |  Rows *must* be in lists.
+---+ z |  Columns *can* be lists, but they have to
| y |   |  be lists of rows.
+---+---+  Columns can also be windows, otherwise there'd
be no leaves in the tree.

So in the above case, the whole thing is a single row R1.
R1 = [ ... ]

Top level rows are always wrapped in a single list.
(You could think of it as a column.)

[
  R1 = [...]
]

Within R1 are 2 columns: C1, and C2

[
  R1 = [
         C1 = [ ... ]
         C2 = z       (leaf)
       ]
]

In order for C1 to have anything in it, it has to have rows.
Within C1 is two rows: R2, and R3

[
  R1 = [
         C1 = [
                R2,
                R3
              ]
         C2 = z       (leaf)
       ]
]

Within R2 and R3 are windows x and y.

[                              +-------------------+ Implied outer column.
  R1 = [                       |R1 +-------------+ | It's just a list of
         C1 = [                |   |C1 +--------+| | Rows. We need some kind
                R2 = [         |   |   |R2 +---+|| | of container for rows.
                       x       |   |   |   | x ||| |
                     ],        |   |   |   |   ||| |
                R3 = [         |   |   |   +---+|| |
                       y       |   |   |R3 +---+|| |
                     ]         |   |   |   | y ||| |
              ],               |   |   |   |   ||| |
         C2 = z       (leaf)   |   |   |   +---+|| |
       ]                       |   |   +--------+| |
]                              |   +-------------+ |
                               +-------------------+

Because R2 only has a single column, x, we can simple
add another row to the column, R4, after R2. If R2 had multiple
columns, then x would need to be wrapped in a column list C3, and
then a row list R4, and then a second row, R5, could be added inside
C3.

This optimization might be a dumb idea, I'm not sure yet.

"Pre-optimization is the root of all evil."

Confused yet? lol