Fixed split_windows function
This included writing a custom dbg tracing function to render #window{}
objects as IDs. This made it much easier, along with the file names,
line numbers, and indentation, to understand what was going on. One
issue was that when I switched to processing a row (i.e. a list of
windows and/or columns) I passed the rows accumulator. This ended up
duplicating a row.