Fix window split and rotate
Also, more work on the dbg function to print out window split tree
shapes during testing. I split out the function-specific parts so I
could use it without both the window rotate and split functions.

Once a window gets to the end of a column, or is on the bottom row, it
can no longer be rotated right/down (clockwise).