Read chars with NIF
The goal here is to read all input as characters, and Erlang can't do
that, even in "raw" mode. Things like Ctrl-c still come across as
interrupts (or something) to the Erlang shell. See notes.

Create a C shared object file that hooks up to a NIF erl module.
Add compile hooks to compile and package up the C.
Remove Erlang interactive shell since it listens for characters and
competes with the NIF.

I have the NIF waiting for 100 ms for a character, but I'm going to
change that to just return 0 if there are no characters, and having
Erlang manage when to try and read: that way the Erlang schedule can do
other things during that time.

I should eventually look into polling to see if it's NIF-compatible, or
even an Erlang port.