Send raw text to screen
When we get text that doesn't parse as a command we send it to the
screen so the screen can send it to the focused window.

When print debug text to the screen, always put the cursor at 5,5 first
so it's always in the same place and doesn't keep filling up the screen.

Beware of mixing up fun cursor_pos/2 and fun cursor_pos_/2. :)