Add on-screen debugging
Print:
- what the PTY receives in readable text
- what cs_io does with what it gets from the PTY.
- any textarea or screen size events
- any unrecognized casts

I'm trying to work on stopping text from overflowing the window but I
needed more clarity in what each different process was seeing:
- the PTY
- cs_io
- window

Now PTY and cs_io have their own dedicated output screen space with
field names. Each new event clears the rest of the line.

For the PTY I make sure to clear the remaining line space the first time
we run out of characters in the PTY buffer; e.g. after receiving an
escape code.

I had to figure out how to parse characters that I wanted to display but
not escape code characters that needed to be sent to the termainal
(iTerm2 in this case).