Don't reset focused window for second row
Each row needs to save it's "primary" window in the state so that the
next row, if it contains the focused window, can move the focus to it.
If the column deletes the "window to move focus to" then we lose the
ability to focus on it.

e.g.

 X    1. do row X
---   2. store "X is window to focus on next"
 Y*   3. do row Y - do not delete "X is the window to focus on next"
      4. Y is focused, move focus to X

The first row in every column is special because it might be focused and
need to know the primary window in the row holding the column:

Here there are 2 outside rows. Row 2 has 2 columns (2,3),
column 2 has two rows (3,4). Y is focused. Y is the top row in
column 2 and needs to move to focus to X, which is in the earlier
sibling (row 1) of the parent row (3) of Y.

[[X],[[[Y],[A]],Z]

Row:
1           X
      -------------  1. Store X as "parent"
  3      Y*  |  Z    2. Do row 2
2     -------|       3. Do col 2
  4      A   |       4. Do row 1
                     5. Y is focused, no "window", use "parent"
Col:     2      3
             1       Column 1 has everything in it.