WIP: adding dispatch and commands
If we're in command mode then we need to dispatch input keys as
commands. I imagine the command mode will be like a buffer that keeps
things in it's memory until enter is pressed.

If we're in insert mode then we need to send the input text to cs_screen
so it can dispatch it to the right buffer, which will then tell the
window what to update. I'm not sure what "what to update" will look like
just yet.

Move the "quit" responsibility to a dedicated "quit" gen_server so that
it's safe from crashing because of any cs_io logic. All it has is a quit
function and it only ever gets one message: 'quit'.

So cs_io -> dispatch -> screen
                     -> command

All of this is in support of opening up a buffer ... I think. I can't
quite remember now.