WIP: add command window
TODO: move to bottom of screen.

The command window starts out as one line but it's a regular window that
can be expanded. You can't change focus to it with the Ctrl+{h,j,k,l}
keys; you have to hit ':' while not in insert mode to focus on the
command window.