Enter/exit insert mode; mapping
Treat <Esc> as special when in insert mode so we can exit insert mode.

Cancel pending erlang:send_after/3 timeout references when we receive
the next possible character in an escape sequence. Note that
erlang:send_after/x and timer:send_after/x use different references and
are not compatible.

Introduce a map lookup server in between cs_io and cs_dispatch to track
whether incoming keys are mapped to another combination of keys. I'm
following Neovim where the first character that's entered is displayed
over top of the character after the cursor, but subsequent potential map
characters are not displayed. After 500 ms Neovim pops off the first
character in the potential map queue and, if in insert mode, inserts it;
however, the line in the window doesn't get updated until the potential
mapping is resolved: i.e. all characters have been popped or a map has
been found.

  I created a mapping of "kkkkkk" -> <Esc> and was able to enter 5 k's
  and then watched as they slowly popped off the potential map queue.
  While there's still k's on the queue, you can add more. It just
  becomes delayed text-entry. If you enter the k's fast enough to catch
  up and get the sixth one entered, then the mapping matches and you
  send <Esc> and exit insert mode.